Move VCard in own component
